+++ example/android/app/src/main/java/com/social/flutteranimationplayground/MainActivity.java
@@ -1,0 +1,6 @@
+package com.social.flutteranimationplayground;
+
+import io.flutter.embedding.android.FlutterActivity;
+
+public class MainActivity extends FlutterActivity {
+}