+++ lib/animated_progress_button.dart
@@ -3,31 +3,46 @@ library animated_progress_button;
 import 'package:flutter/material.dart';
 
 class AnimatedButton extends StatefulWidget {
+  const AnimatedButton(
+      {Key key,
+        @required this.color,
+        @required this.text,
+        @required this.controller,
+        this.loadedIcon,
+        this.height,
+        this.loadingText,
+        this.onPressed})
+      : super(key: key);
+
+  /// Color of the button
   final Color color;
+
+  /// Initial text display
   final String text;
+
+  /// Loading text
   final String loadingText;
+
+  /// You can also customized the last icon appear in the animation progress.
   final Widget loadedIcon;
+
+  /// Default 50.
   final double height;
+
+  /// AnimatedButtonController to help control the animation progress.
+  /// Call [controller.completed()] when getting response from your request
+  /// Call [controller.reset()] to restart button animation
   final AnimatedButtonController controller;
+
+  /// Customize on pressing action
   final VoidCallback onPressed;
-
-
-  const AnimatedButton({
-    Key key,
-    @required this.color,
-    @required this.text,
-    @required this.controller,
-    this.loadedIcon,
-    this.height,
-    this.loadingText,
-    this.onPressed}) : super(key: key);
-
 
   @override
   _AnimatedButtonState createState() => _AnimatedButtonState();
 }
 
-class _AnimatedButtonState extends State<AnimatedButton> with TickerProviderStateMixin {
+class _AnimatedButtonState extends State<AnimatedButton>
+    with TickerProviderStateMixin {
   static int simulatedLoadingTime = 3;
   bool _isLoading = false;
   AnimationController _animationController;
@@ -46,16 +61,28 @@ class _AnimatedButtonState extends State<AnimatedButton> with TickerProviderStat
       }
       _animationController.forward(from: 0.6);
     });
-    _animationController = AnimationController(duration: Duration(seconds: simulatedLoadingTime), vsync: this)
+    _animationController = AnimationController(
+        duration: Duration(seconds: simulatedLoadingTime), vsync: this)
       ..addListener(() {
-        if (_animationController.value >= 0.6 && !(widget.controller?.value ?? true)) {
+        if (_animationController.value >= 0.6 &&
+            !(widget.controller?.value ?? true)) {
           _animationController.stop(canceled: false);
         }
         setState(() {});
       });
-    _progressAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(CurvedAnimation(parent: _animationController, curve: Interval(0.0, 0.7, curve: Curves.ease)));
-    _buttonWidthAnimation = Tween<double>(begin: 400, end: 50).animate(CurvedAnimation(parent: _animationController, curve: Interval(0.7, 0.9, curve: Curves.decelerate)));
-    _checkIconOffsetAnimation = Tween<Offset>(begin: Offset(0, -2), end: Offset.zero).animate(CurvedAnimation(parent: _animationController, curve: Interval(0.9, 1.0, curve: Curves.easeOutBack)));
+    _progressAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(
+        CurvedAnimation(
+            parent: _animationController,
+            curve: Interval(0.0, 0.7, curve: Curves.ease)));
+    _buttonWidthAnimation = Tween<double>(begin: 400, end: 50).animate(
+        CurvedAnimation(
+            parent: _animationController,
+            curve: Interval(0.7, 0.9, curve: Curves.decelerate)));
+    _checkIconOffsetAnimation =
+        Tween<Offset>(begin: Offset(0, -2), end: Offset.zero).animate(
+            CurvedAnimation(
+                parent: _animationController,
+                curve: Interval(0.9, 1.0, curve: Curves.easeOutBack)));
   }
 
   bool get _progressLoadingCompleted => _progressAnimation.value == 1.0;
@@ -86,10 +113,8 @@ class _AnimatedButtonState extends State<AnimatedButton> with TickerProviderStat
                       offset: Offset(0, 3),
                       color: Colors.black12,
                       blurRadius: 10.0,
-                      spreadRadius: 1.0
-                  )
-                ]
-            ),
+                      spreadRadius: 1.0)
+                ]),
             child: Stack(
               children: <Widget>[
                 Row(
@@ -97,27 +122,37 @@ class _AnimatedButtonState extends State<AnimatedButton> with TickerProviderStat
                   children: <Widget>[
                     Padding(
                       padding: const EdgeInsets.all(16.0),
-                      child: Text(_progressLoadingCompleted ? "" : _isLoading ? (widget.loadingText ?? 'In progress ...') : (widget.text ?? ''),
+                      child: Text(
+                          _progressLoadingCompleted
+                              ? ""
+                              : _isLoading
+                                  ? (widget.loadingText ?? 'In progress ...')
+                                  : (widget.text ?? ''),
                           style: TextStyle(color: Colors.white, fontSize: 15)),
                     ),
                   ],
                 ),
-                _progressLoadingCompleted ? Container() : Container(
-                  height: 50,
-                  child: ClipRRect(
-                    borderRadius: BorderRadius.all(Radius.circular(32.0)),
-                    child: LinearProgressIndicator(
-                      value: _progressAnimation.value,
-                      valueColor: AlwaysStoppedAnimation<Color>(Colors.white.withOpacity(.5)),
-                      backgroundColor: Colors.transparent,
-                    ),
-                  ),
-                ),
+                _progressLoadingCompleted
+                    ? Container()
+                    : Container(
+                        height: 50,
+                        child: ClipRRect(
+                          borderRadius: BorderRadius.all(Radius.circular(32.0)),
+                          child: LinearProgressIndicator(
+                            value: _progressAnimation.value,
+                            valueColor: AlwaysStoppedAnimation<Color>(
+                                Colors.white.withOpacity(.5)),
+                            backgroundColor: Colors.transparent,
+                          ),
+                        ),
+                      ),
                 Visibility(
                   visible: _visibleIcon,
                   child: SlideTransition(
                       position: _checkIconOffsetAnimation,
-                      child: Center(child: widget.loadedIcon ?? Icon(Icons.check, color: Colors.white))),
+                      child: Center(
+                          child: widget.loadedIcon ??
+                              Icon(Icons.check, color: Colors.white))),
                 )
               ],
             )),
@@ -151,5 +186,3 @@ class AnimatedButtonController extends ValueNotifier<bool> {
     }
   }
 }
-
-